requirements.txt
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.5
cramjam==2.5.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
fastparquet==0.8.3
fsspec==2022.10.0
greenlet==1.1.3.post0
ipykernel==6.16.1
ipython==8.5.0
jedi==0.18.1
jupyter_client==7.4.3
jupyter_core==4.11.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.3
pure-eval==0.2.2
Pygments==2.13.0
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.5
pywin32==304
pyzmq==24.0.1
six==1.16.0
SQLAlchemy==1.4.42
stack-data==0.5.1
tornado==6.2
traitlets==5.5.0
wcwidth==0.2.5
